ha_nostr_alert_addon/requirements.txt
nostr-sdk==0.33.0
flask==3.0.0
pyyaml==6.0.1
